refactor: move fee input to review step - username registration
