+++ src/domains/transaction/components/ValidatorRegistrationForm/ReviewStep.tsx
@@ -72,7 +72,7 @@ export const ReviewStep = ({
 								{t("TRANSACTION.VALIDATOR_PUBLIC_KEY")}
 							</DetailTitle>
 							<div className="no-ligatures truncate text-sm font-semibold leading-[17px] text-theme-secondary-900 dark:text-theme-secondary-200 sm:text-base sm:leading-5">
-								{getValues("validatorPublicKey")}
+								{validatorPublicKey}
 							</div>
 						</div>
 					</div>
